Fix server rendering error where checksums for client & server don't match.
The crux here is that calling the spread operator on the user object yields
different results on client compared to the server. Not sure why this is - my
hunch is that this is because the `req.user` object from expressjs is a complex
structure, and thus can't be handled?

As a side effect, I made the Main component become the only controller-view,
instead of plumbing props all the way down - the container elements can be
dumb.

Also, upgrade to React v13.